# Conflicts:
#	src/ailego/buffer/vector_page_table.cc
#	src/core/algorithm/diskann/diskann_context.cc
#	src/core/algorithm/diskann/diskann_file_reader.cc
#	src/core/algorithm/diskann/diskann_file_reader.h
#	src/core/algorithm/diskann/diskann_indexer.cc
#	src/include/zvec/ailego/buffer/vector_page_table.h
#	tests/ailego/CMakeLists.txt
#	tests/ailego/buffer/vector_page_table_test.cc
#	tests/ailego/io/iouring_loader_test.cc
#	tests/core/algorithm/diskann/diskann_builder_test.cc
#	tests/core/algorithm/diskann/diskann_file_reader_aio_test.cc
#	tests/core/algorithm/diskann/diskann_file_reader_test.cc
#	tests/core/utility/visit_filter_test.cc
#	tools/core/local_builder.cc

+++ examples/c/optimized_example.c
@@ -43,7 +43,7 @@ static float *create_test_vector(size_t dimension) {
   }
 
   for (size_t i = 0; i < dimension; i++) {
-    vector[i] = (float)rand() / RAND_MAX;
+    vector[i] = (float)rand() / (float)RAND_MAX;
   }
 
   return vector;
